01:49 AM 11/28/2023

+++ d1-master/Project_Economic/src/main/java/com/example/project_economic/dto/ProductDto.java
@@ -21,5 +21,5 @@ public class ProductDto {
     private Long  categoryId;
     private Boolean is_deteted;
     private Boolean is_actived;
-    private Integer sellerId;
+    private Long sellerId;
 }
